move BeHuman project files into new BeHuman folder
