+++ src/components/projects/BeHuman/BeHumanDetails.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { Link } from 'react-router-dom';
-import ProjectsDetails from '../continents/ProjectsDetails';
-import beHumanIcon from '../../assets/beHuman_halfSize.png';
+import ProjectsDetails from '../../continents/ProjectsDetails';
+import beHumanIcon from '../../../assets/beHuman_halfSize.png';
 import styles from './BeHumanDetails.module.css';
 
 const BeHumanDetails = () => {
